+++ tapo/src/responses/child_device_list_hub_result/ke100_result.rs
@@ -37,6 +37,7 @@ pub struct KE100Result {
     pub jamming_rssi: i16,
     pub jamming_signal_level: u8,
     pub mac: String,
+    pub model: String,
     pub nickname: String,
     pub oem_id: String,
     pub parent_device_id: String,