+++ tapo/src/responses/child_device_list_hub_result/s200_result.rs
@@ -22,6 +22,7 @@ pub struct S200Result {
     pub jamming_rssi: i16,
     pub jamming_signal_level: u8,
     pub mac: String,
+    pub model: String,
     pub nickname: String,
     pub oem_id: String,
     pub parent_device_id: String,