+++ tapo/src/responses/child_device_list_hub_result/t31x_result.rs
@@ -39,6 +39,7 @@ pub struct T31XResult {
     pub jamming_rssi: i16,
     pub jamming_signal_level: u8,
     pub mac: String,
+    pub model: String,
     pub nickname: String,
     pub oem_id: String,
     pub parent_device_id: String,